%% File Name: data.kess.Rd
%% File Version: 0.02

\name{data.kess}
\alias{data.kess}
\alias{data.kess.math}
\alias{data.kess.read}
\alias{data.kess.sci}
\alias{data.kess.figIntell}

\docType{data}


\title{
Datasets for Longitudinal KESS Study
}

\description{
These are synthetic datasets for the longitudinal German KESS study in the domains
mathematics, reading, science and figural intelligence and Grades 4, 7, 8 and 10.
School track (with levels 1, 2 and 3) and gender variable female are also
included as demographic variables.
}

\usage{
data(data.kess.math)
data(data.kess.read)
data(data.kess.sci)
data(data.kess.figIntell)
}


\format{
The datasets for each of the four domains are included in a list where
each list entry corresponds to one grade.
}



%% \keyword{datasets}
