\name{data.long}
\alias{data.long}
\docType{data}
\title{
Longitudinal Dataset
}
\description{
This dataset contains 200 observations on
12 items. 6 items (\code{I1T1},... ,\code{I6T1})
were administered at measurement occasion T1
and 6 items at T2 (\code{I3T2},... , \code{I8T2}).
The first column contains the student identifier.
}
\usage{data(data.long)}
%\examples{
%# longitudinal scaling Rasch model and 2PL model
%# rasch.mml2,  smirt, # TAM, # mirt # lme4 # MCMCglmm
%}
\keyword{datasets}
