\name{data.ml}
\alias{data.ml}
\alias{data.ml1}
\alias{data.ml2}

\docType{data}
\title{
Multilevel Datasets
}
\description{
Datasets for conducting multilevel IRT analysis. This dataset is used
in the examples of the function \code{\link{mcmc.2pno.ml}}. 
}
\usage{
data(data.ml1)
data(data.ml2)
}
\format{
  \code{data.ml1} \cr
  A data frame with 2000 student observations in 100 classes on 17 variables.
  The first variable \code{group} contains the class identifier.
  The remaining 16 variables are dichotomous test items. \cr 
  
  \code{data.ml2} \cr
  A data frame with 2000 student observations in 100 classes on 6 variables.
  The first variable \code{group} contains the class identifier.
  The remaining 5 variables are polytomous test items. %%\cr \cr  
}
\keyword{datasets}
