\name{data.ml1}
\alias{data.ml1}
\docType{data}
\title{
Dataset Multilevel 1
}
\description{
Dataset for conducting multilevel IRT analysis. This dataset is used
in the examples of the function \code{\link{mcmc.2pno.ml}}. 
}
\usage{data(data.ml1)}
\format{
  A data frame with 2000 student observations in 100 classes on 17 variables.
  The first variable \code{group} contains the class identifier.
  The remaining 16 variables are dichotomous test items.
}
\keyword{datasets}
