\name{data.ml2}
\alias{data.ml2}
\docType{data}
\title{
Dataset Multilevel 2
}
\description{
Dataset for conducting multilevel random item analysis. This dataset is used
in the examples of the function \code{\link{mcmc.2pno.ml}}. 
}
\usage{data(data.ml2)}
\format{
  A data frame with 2000 student observations in 100 classes on 6 variables.
  The first variable \code{group} contains the class identifier.
  The remaining 5 variables are dichotomous test items.
}
\keyword{datasets}
