\name{data.ratings1}
\alias{data.ratings1}
\docType{data}
\title{
Rating Dataset 1
}
\description{
This data set contains ratings for 135 students from
7 raters on 5 items.
}
\usage{data(data.ratings1)}
\format{
  A data frame with 274 observations on the following 7 variables.
  \describe{
    \item{\code{idstud}}{Student identifier}
    \item{\code{rater}}{Rater identifier}
    \item{\code{k1}}{Item 1}
    \item{\code{k2}}{Item 2}
    \item{\code{k3}}{Item 3}
    \item{\code{k4}}{Item 4}
    \item{\code{k5}}{Item 5}
  }
}

\source{
Data from a 2009 Austrian survey of national educational
standards for 8th graders in German language writing.
Variables \code{k1} to \code{k5} denote several rating
criteria of writing compentency.
}


\section{Example Index}{
\code{\link{rm.facets}} (Example 1, Example 2),
\code{\link{rm.hrm}} (Example 1),
}

%\examples{
%# include an example of rating data here	
%}	

\keyword{datasets}
