% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_3pno_testlet_draw_itempars.R
\name{mcmc_3pno_testlet_draw_itempars}
\alias{mcmc_3pno_testlet_draw_itempars}
\title{'''''''''''''''''''''''''''''''''''''''
'''''''''''''''''''''''''''''''''''''''}
\usage{
mcmc_3pno_testlet_draw_itempars(theta, Z, I, N, weights, gamma.testlet,
  testletgroups, param, TT, a.testletM)
}
\description{
'''''''''''''''''''''''''''''''''''''''
'''''''''''''''''''''''''''''''''''''''
}
