%% File Name: sirt-defunct.Rd
%% File Version: 0.08

\name{sirt-defunct}
\alias{sirt-defunct}
\alias{rasch.conquest}
\alias{rasch.pml2}
\alias{yen.q3}
\alias{testlet.yen.q3}



\title{Defunct \pkg{sirt} Functions}

\description{
These functions have been removed or replaced in the \pkg{sirt}
package.
}

\usage{
rasch.conquest(...)
rasch.pml2(...)
testlet.yen.q3(...)
yen.q3(...)
}

\arguments{
    \item{\dots}{Arguments to be passed.}
}

\details{
The \code{rasch.conquest} function has been replaced by
\code{\link{R2conquest}}.

The \code{rasch.pml2} function has been superseded by
\code{\link{rasch.pml3}}.

The \code{testlet.yen.q3} function has been replaced by
\code{\link{Q3.testlet}}.

The \code{yen.q3} function has been replaced by
\code{\link{Q3}}.
}

% \examples{
% #############################################################################
% }
