\name{summary.gom}
\alias{summary.gom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{gom}
}
\description{
S3 method to summarize objects of class \code{gom}.
}
\usage{
\method{summary}{gom}(object,...)	
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{gom}
}
  \item{\dots}{
Further arguments to be passed
}
}
\author{
Alexander Robitzsch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
For examples see \code{\link{gom.em}}.
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{summary}
