\name{summary.lsdm}
\alias{summary.lsdm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{lsdm}
}
\description{
S3 method to summarize objects of class \code{lsdm}.
}
\usage{
\method{summary}{lsdm}(object,...)	
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{lsdm}
}
  \item{\dots}{
Further arguments to be passed
}
}

\author{
Alexander Robitzsch
}

\seealso{
\code{\link{lsdm}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%% \keyword{Least Squares Distance Method (LSDM)}
%% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{summary}
