%% File Name: summary.mcmc.sirt.Rd
%% File Version: 0.08

\name{summary.mcmc.sirt}
\alias{summary.mcmc.sirt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{mcmc.sirt}
}
\description{
S3 method to summarize objects of class \code{mcmc.sirt}.
This object is generated by following functions: \code{\link{mcmc.2pno}},
\code{\link{mcmc.2pnoh}}, \code{\link{mcmc.3pno.testlet}},
\code{\link{mcmc.2pno.ml}}
}
\usage{
\method{summary}{mcmc.sirt}(object,digits=3, file=NULL, ...)
}


\arguments{
  \item{object}{
Object of class \code{mcmc.sirt}
}
\item{digits}{Number of digits after decimal}
\item{file}{Optional file name to which \code{summary} output is written}
  \item{\dots}{
Further arguments to be passed
}
}
\author{
Alexander Robitzsch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{mcmc.2pno}}, \code{\link{mcmc.2pnoh}},
\code{\link{mcmc.3pno.testlet}}, \code{\link{mcmc.2pno.ml}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{summary}
