\name{summary.prob.guttman}
\alias{summary.prob.guttman}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{prob.guttman}
}
\description{
S3 method to summarize objects of class \code{prob.guttman}.
}
\usage{
\method{summary}{prob.guttman}(object,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{prob.guttman}
}
  \item{\dots}{
Further arguments to be passed
}
}

\author{
Alexander Robitzsch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
See \code{\link{prob.guttman}} for examples.
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%% \keyword{ ~kwd1 }
\keyword{summary}

