\name{summary.rasch.copula2}
\alias{summary.rasch.copula2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{rasch.copula2}
}
\description{
S3 method to summarize objects of class \code{rasch.copula2}.
}
\usage{
\method{summary}{rasch.copula2}(object,...)	
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{rasch.copula2}
}
  \item{\dots}{
Further arguments to be passed
}
}


\author{
Alexander Robitzsch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{rasch.copula2}}
    }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%% \keyword{ ~kwd1 }
\keyword{summary}