\name{summary.rasch.jml}
\alias{summary.rasch.jml}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{rasch.jml}
}
\description{
S3 method to summarize objects of class \code{rasch.jml}.
}
\usage{
\method{summary}{rasch.jml}(object,...)	
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{rasch.jml}
}
  \item{\dots}{
Further arguments to be passed
}
}
\author{
Alexander Robitzsch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{rasch.jml}}
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%% \keyword{Joint Maximum Likelihood Estimation}
%% \keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{summary}
