\name{summary.rasch.mirtlc}
\alias{summary.rasch.mirtlc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{rasch.mirtlc}
}
\description{
Summary Method for Objects of Class \code{rasch.mirtlc}
}
\usage{
\method{summary}{rasch.mirtlc}(object,...)	
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{rasch.mirtlc}
}
  \item{\dots}{
Further arguments to be passed
}
}

\author{
Alexander Robitzsch
}

\seealso{
For examples see \code{\link{rasch.mirtlc}}.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%%\keyword{ ~kwd1 }
%%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{summary}
