\name{summary.rm.facets}
\alias{summary.rm.facets}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{rm.facets}
}
\description{
S3 method to summarize objects of class \code{rm.facets}.
}
\usage{
\method{summary}{rm.facets}(object,...)	
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{rm.facets}
}
  \item{\dots}{
Further arguments to be passed
}
}
\author{
Alexander Robitzsch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{rm.facets}}
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{summary}
