\name{summary.smirt}
\alias{summary.smirt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Summary Method for Objects of Class \code{smirt}
}
\description{
S3 method to summarize objects of class \code{smirt}.
}
\usage{
\method{summary}{smirt}(object,...)	
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class \code{smirt}
}
  \item{\dots}{
Further arguments to be passed
}
}

\author{
Alexander Robitzsch
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
For examples see \code{\link{smirt}}.
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
\keyword{summary}
