%% File Name: data.pisa2006Read.Rd
%% File Version: 0.02

\name{data.pisa2006Read}
\alias{data.pisa2006Read}
\docType{data}
\title{
Dataset PISA 2006 Reading
}
\description{
This is the PISA 2006 dataset of item responses from 26 OECD countries.
}
\usage{
data(data.pisa2006Read)
}


\format{
The dataset includes item responses, student weights and their respective 
replicate weights.
}



%% \keyword{datasets}
